requirements.txt
langchain 
langgraph 
langchain-ollama
langchain_community
redis
fastapi
uvicorn[standard]
celery
GitPython
PyGithub
